requirements.txt
requests
pyyaml
prefect
pandas
seaborn
category_encoders
ipywidgets
scikit-learn
plotly